Persistent player's stuffs

+++ src/GameState/StateSplash.cpp
@@ -14,7 +14,8 @@ void StateSplash::init()
 	m_timer.restart();
 
 	m_sfmlLogo.setTexture(*TextureCache::Get().getTexture("sfml-logo-small.png"));
-	m_sfmlLogo.setOrigin(m_sfmlLogo.getLocalBounds().width/2, m_sfmlLogo.getLocalBounds().height/2);
+	m_sfmlLogo.setOrigin(static_cast<int>(m_sfmlLogo.getLocalBounds().width/2),
+						 static_cast<int>(m_sfmlLogo.getLocalBounds().height/2));
 	m_sfmlLogo.setPosition({Screen::Get().halfWidth, Screen::Get().halfHeight -24});
 
 	Renderer::Get().setCameraPos({Screen::Get().halfWidth, Screen::Get().halfHeight});
